Add a 'subscriptions' module as well.
